+++ diskette/factories/__init__.py
@@ -1,6 +1,10 @@
+from .dump import DumpFileFactory
+from .key import APIkeyFactory
 from .user import UserFactory
 
 
 __all__ = [
+    "APIkeyFactory",
+    "DumpFileFactory",
     "UserFactory",
 ]